snd_codec.c (S_CodecUtilOpen): fix again after rev.410.
git-svn-id: svn://svn.code.sf.net/p/quakespasm/code/trunk/quakespasm@411 af15c1b1-3010-417e-b628-4374ebc0bcbd

+++ Quake/snd_codec.c
@@ -232,7 +232,7 @@ snd_stream_t *S_CodecUtilOpen(const char *filename, snd_codec_t *codec)
 	/* Try to open the file */
 	length = (long) COM_FOpenFile(filename, &handle, NULL);
 	pak = file_from_pak;
-	if (length == (size_t)-1)
+	if (length == -1)
 	{
 		Con_DPrintf("Couldn't open %s\n", filename);
 		return NULL;
